add react to SystemBrowser sample and removed ElectronReactSample
