--- samples/msal-node-samples/ElectronReactTestApp/.gitignore
@@ -1,1 +1,0 @@
-.webpack
