--- samples/msal-node-samples/ElectronReactTestApp/data/cacheTemplate.json
@@ -1,7 +1,0 @@
-{
-    "Account": {},
-    "IdToken": {},
-    "AccessToken": {},
-    "RefreshToken": {},
-    "AppMetadata": {}
-}
