--- samples/msal-node-samples/ElectronReactTestApp/package.json
@@ -1,111 +1,0 @@
-{
-  "name": "electron-react-test-app",
-  "productName": "electron-react-test-app",
-  "version": "1.0.0",
-  "description": "A sample Electron and React application that uses msal-node to acquire tokens",
-  "main": ".webpack/main",
-  "scripts": {
-    "build": "tsc",
-    "start": "electron-forge start",
-    "package": "electron-forge package",
-    "build:package": "cd ../../../lib/msal-common && npm run build && cd ../msal-node && npm run build",
-    "start:build": "npm run build:package && npm start",
-    "lint": "eslint --ext .ts,.tsx ."
-  },
-  "devDependencies": {
-    "@electron-forge/cli": "^6.0.0-beta.63",
-    "@electron-forge/maker-deb": "^6.0.0-beta.63",
-    "@electron-forge/maker-rpm": "^6.0.0-beta.63",
-    "@electron-forge/maker-squirrel": "^6.0.0-beta.63",
-    "@electron-forge/maker-zip": "^6.0.0-beta.63",
-    "@electron-forge/plugin-webpack": "^6.0.0-beta.63",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.22",
-    "@typescript-eslint/eslint-plugin": "^5.9.1",
-    "@typescript-eslint/parser": "^5.9.1",
-    "@vercel/webpack-asset-relocator-loader": "^1.7.0",
-    "css-loader": "^6.5.1",
-    "electron": "^18.2.3",
-    "fork-ts-checker-webpack-plugin": "^6.5.0",
-    "node-loader": "^2.0.0",
-    "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.6",
-    "typescript": "~4.5.4"
-  },
-  "dependencies": {
-    "@azure/msal-node": "^1.8.0",
-    "axios": "^0.27.2",
-    "bootstrap": "^4.5.3",
-    "dotenv": "^16.0.0",
-    "electron-fetch": "^1.7.4",
-    "electron-squirrel-startup": "^1.0.0",
-    "history": "^5.2.0",
-    "install": "^0.13.0",
-    "npm": "^8.5.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-router-dom": "^6.3.0"
-  },
-  "keywords": [
-    "msal-node",
-    "electron",
-    "react",
-    "typescript"
-  ],
-  "license": "MIT",
-  "config": {
-    "forge": {
-      "packagerConfig": {},
-      "makers": [
-        {
-          "name": "@electron-forge/maker-squirrel",
-          "config": {
-            "name": "electron-react-test-app"
-          }
-        },
-        {
-          "name": "@electron-forge/maker-zip",
-          "platforms": [
-            "darwin"
-          ]
-        },
-        {
-          "name": "@electron-forge/maker-deb",
-          "config": {}
-        },
-        {
-          "name": "@electron-forge/maker-rpm",
-          "config": {}
-        }
-      ],
-      "plugins": [
-        [
-          "@electron-forge/plugin-webpack",
-          {
-            "mainConfig": "./webpack.main.config.js",
-            "renderer": {
-              "config": "./webpack.renderer.config.js",
-              "entryPoints": [
-                {
-                  "html": "./src/index.html",
-                  "js": "./src/renderer.ts",
-                  "name": "main_window",
-                  "preload": {
-                    "js": "./src/preload.ts"
-                  }
-                }
-              ]
-            }
-          }
-        ]
-      ]
-    }
-  }
-}
