--- samples/msal-node-samples/ElectronReactTestApp/src/App.ts
@@ -1,2 +1,0 @@
-import Main from './Main';
-Main.main();
