--- samples/msal-node-samples/ElectronReactTestApp/src/GraphReponseTypes.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-/**
- * Graph data about the user.
- */
-export type UserInfo = {
-    businessPhones?: Array<string>;
-    displayName?: string;
-    givenName?: string;
-    id?: string;
-    jobTitle?: string;
-    mail?: string;
-    mobilePhone?: string;
-    officeLocation?: string;
-    preferredLanguage?: string;
-    surname?: string;
-    userPrincipalName?: string;
-};
