--- samples/msal-node-samples/ElectronReactTestApp/src/app/index.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import App from './App';
-import { HashRouter } from 'react-router-dom';
-
-import 'bootstrap/dist/css/bootstrap.min.css';
-
-ReactDOM.render(
-    <React.StrictMode>
-        <HashRouter>
-            <App />
-        </HashRouter>
-    </React.StrictMode>,
-    document.getElementById('root')
-);
