--- samples/msal-node-samples/ElectronReactTestApp/src/app/pages/Home.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const Home = () => {
-    return (
-        <div>
-            <h5 className="text-center">Microsoft Authentication Library For Node Electron Sample</h5>
-        </div>
-    );
-};
