--- samples/msal-node-samples/ElectronReactTestApp/src/app/styles/App.css
@@ -1,22 +1,0 @@
-footer {
-    position: fixed;
-    bottom: 5%;
-    width: 100%;
-}
-
-.footer-text {
-    font-size: small;
-    text-align: center;
-    flex: 1 1 auto;
-}
-
-.App {
-    text-align: center;
-}
-
-.data-area-div {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly
-}
