--- samples/msal-node-samples/ElectronReactTestApp/src/app/styles/Nav.css
@@ -1,11 +1,0 @@
-.test-link:link {
-    color: #ffff
-}
-
-.test-link:visited {
-    color: #ffff
-}
-
-.test-link:active {
-    color: blue;
-}
