--- samples/msal-node-samples/ElectronReactTestApp/src/index.css
@@ -1,4 +1,0 @@
-body {
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
-    margin: auto;
-}
