--- samples/msal-node-samples/ElectronReactTestApp/src/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>MSAL Node Electron with React Sample App</title>
-</head>
-
-<body>
-    <div id="root"></div>
-</body>
-
-</html>
