--- samples/msal-node-samples/ElectronReactTestApp/webpack.plugins.js
@@ -1,4 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
-
-module.exports = [new ForkTsCheckerWebpackPlugin()];
