--- samples/msal-node-samples/ElectronReactTestApp/webpack.renderer.config.js
@@ -1,18 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const rules = require('./webpack.rules');
-const plugins = require('./webpack.plugins');
-
-rules.push({
-  test: /\.css$/,
-  use: [{ loader: 'style-loader' }, { loader: 'css-loader' }],
-});
-
-module.exports = {
-  module: {
-    rules,
-  },
-  plugins: plugins,
-  resolve: {
-    extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
-  },
-};
