+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/README.md
@@ -196,5 +196,4 @@ private static onOpenUrl(event: any): void {
 1. The app will open a browser to Azure AD for the user to sign-in.
 ![Screenshot](./ReadmeFiles/Screenshot-browser.png)
 1. Select the **Profile** button on the navigation bar. This will make a call to the Microsoft Graph API.
-1. Select the **Mail** button on the navigation bar. This will make a call to the Microsoft Graph API.
 ![Screenshot](./ReadmeFiles/Screenshot-profile.png)
