+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/src/Constants.ts
@@ -19,6 +19,7 @@ export enum IpcMessages {
     GET_MAIL = "GET_MAIL",
     SET_MAIL = "SET_MAIL",
     GET_AUTH_CODE_URL = "GET_AUTH_CODE_URL",
+    GET_ACCOUNT = "GET_ACCOUNT",
 }
 
 export const APPLICATION_DIMENSIONS = {
