+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/src/Renderer.ts
@@ -2,31 +2,23 @@
 // Licensed under the MIT License
 
 /**
- * The renderer API is exposed by the preload script found in the preload.ts
- * file in order to give the renderer access to the Node API in a secure and
- * controlled way
+ * This file will automatically be loaded by webpack and run in the "renderer" context.
+ * To learn more about the differences between the "main" and the "renderer" context in
+ * Electron, visit:
+ *
+ * https://electronjs.org/docs/tutorial/application-architecture#main-and-renderer-processes
+ *
+ * By default, Node.js integration in this file is disabled. When enabling Node.js integration
+ * in a renderer process, please be aware of potential security implications. You can read
+ * more about security risks here:
+ *
+ * https://electronjs.org/docs/tutorial/security
+ *
+ * To enable Node.js integration in this file, open up `main.js` and enable the `nodeIntegration`
+ * flag
  */
 
-import "bootstrap";
-import "./index.scss";
+import "./app/index.tsx";
+import "./index.css";
 
-declare const window: any;
-
-window.api.startUiManager();
-// UI event handlers
-document.querySelector("#SignIn").addEventListener("click", () => {
-    window.api.sendLoginMessage();
-});
-
-document.querySelector("#SignOut").addEventListener("click", () => {
-    window.api.sendSignoutMessage();
-});
-
-document.querySelector("#seeProfile").addEventListener("click", () => {
-    window.api.sendSeeProfileMessage();
-});
-
-document.querySelector("#readMail").addEventListener("click", () => {
-    window.api.sendReadMailMessage();
-});
