+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/src/app/pages/Home.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const Home = () => {
+    return (
+        <div>
+            <h5 className="text-center">
+                Microsoft Authentication Library For Node Electron Sample
+            </h5>
+        </div>
+    );
+}
