+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/src/config/AAD.json
@@ -5,7 +5,7 @@
     },
     "resourceApi": {
         "endpoint": "https://graph.microsoft.com/v1.0/me",
-        "scopes": ["User.Read", "Mail.Read"]
+        "scopes": ["User.Read"]
     },
     "customProtocol": {
         "name": "msal4b0db8c2-9f26-4417-8bde-3f0e3656f8e0"
