+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/src/index.css
@@ -1,0 +1,5 @@
+body {
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+        Helvetica, Arial, sans-serif;
+    margin: auto;
+}
