--- samples/msal-node-samples/ElectronSystemBrowserTestApp/src/index.scss
@@ -1,1 +1,0 @@
-@import "~bootstrap/scss/bootstrap";
