+++ samples/msal-node-samples/ElectronSystemBrowserTestApp/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "jsx": "react",
     "allowJs": true,
     "module": "commonjs",
     "skipLibCheck": true,
